backend base

+++ backend/controllers/gameController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler')
 
 const Game = require('../models/gameModel')
+const User = require('../models/userModel')
 
 
 // get games
@@ -8,7 +9,7 @@ const Game = require('../models/gameModel')
 //acess private
 const getGames = asyncHandler(async (req,res) => {
 
-    const games = await Game.find()
+    const games = await Game.find({user: req.user.id})
     res.status(200).json(games)
 })
 
@@ -24,6 +25,7 @@ const setGame = asyncHandler(async (req,res) => {
 
     const game = await Game.create({
         text: req.body.text,
+        user: req.user.id
     })
 
 
@@ -39,6 +41,19 @@ const updateGame = asyncHandler(async (req,res) => {
     if(!game){
         res.status(400)
         throw new Error('Game not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // make sure logged user matches game user
+    if(game.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+
     }
     const updatedGame = await Game.findByIdAndUpdate(req.params.id,req.body,{
         new: true,
@@ -56,6 +71,20 @@ const deleteGame = asyncHandler(async (req,res) => {
         res.status(400)
         throw new Error('Game not found')
     }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // make sure logged user matches game user
+    if(game.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+
+    }
+
     await game.deleteOne()
